backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
google-cloud-firestore
google-cloud-storage
python-dotenv
crewai